Manage University

+++ Homework Java/Gestione Università/src/Student.java
@@ -46,14 +46,14 @@ public class Student implements java.lang.Comparable<Student> {
     }
 
     public double getECTSAverage(){
-        int cfu = 0;
+        double cfu = 0.0;
         double media = 0.0;
         if(exams.isEmpty())
             return 0.0;
         for (AbstractExam e : exams) {
             if (e.getGrade() > 17) {
-                cfu += e.getCredits();
-                media += e.getGrade() * e.getCredits();
+                cfu += (double)e.getCredits();
+                media += (double)e.getGrade() * (double)e.getCredits();
             }
         }
         return media / cfu;
